updated sims

+++ slr_analyses/Experiments/sims_latent.R
@@ -48,7 +48,7 @@ res = foreach(
   # Tuning parameters###########################################################
   
   # Settings to toggle with
-  sigma.settings = "latentVarModel"
+  settings.name = "ContinuousResponse"
   n = 100
   p = 30
   K = 10
@@ -58,8 +58,8 @@ res = foreach(
   tol = 1e-4
   sigma_y = 0.1 # sigma (for y)
   sigma_x = 0.1 # sigma_j (for x)
-  SBP.true = matrix(c(1, 1, 1, -1, -1, -1, rep(0, p - 6)))
-  # SBP.true = matrix(c(1, 1, 1, 1, -1, rep(0, p - 5)))
+  # SBP.true = matrix(c(1, 1, 1, -1, -1, -1, rep(0, p - 6)))
+  SBP.true = matrix(c(1, 1, 1, 1, -1, rep(0, p - 5)))
   ilrtrans.true = getIlrTrans(sbp = SBP.true, detailed = TRUE)
   # ilrtrans.true$ilr.trans = transformation matrix (used to be called U) 
   #   = ilr.const*c(1/k+,1/k+,1/k+,1/k-,1/k-,1/k-,0,...,0)
@@ -70,7 +70,7 @@ res = foreach(
   ulimit = 0.5
   
   file.end = paste0(
-    "_", sigma.settings,
+    "_", settings.name,
     "_", paste0(
       paste(which(SBP.true == 1), collapse = ""), "v", 
       paste(which(SBP.true == -1), collapse = "")),
@@ -87,42 +87,52 @@ res = foreach(
   
   ##############################################################################
   # generate data
-  # get latent variable
-  U.all = matrix(runif(min = -ulimit, max = ulimit, 2 * n), ncol = 1)
-  # simulate y from latent variable
-  y.all = as.vector(b0 + b1 * U.all + rnorm(2 * n) * sigma_y)
-  # simulate X: 
-  epsj.all = matrix(rnorm(2 * n * (p - 1)), nrow = (2 * n)) * sigma_x
-  a1 = theta.value * ilrtrans.true$ilr.trans[-p] 
-  #   alpha1j = {
-  #     c1=theta*ilr.const/k+   if j \in I+
-  #     -c2=-theta*ilr.const/k-  if j \in I-
-  #     0                       o/w
-  #   }
-  alrXj.all = a0 + U.all %*% t(a1) + epsj.all #log(Xj/Xp) =alpha0j+alpha1j*U+epsj
-  X.all <- alrinv(alrXj.all)
-  colnames(X.all) = paste0('s', 1:p)
-  
-  # subset out training and test sets
-  X = X.all[1:n, ]
-  X.test = X.all[-(1:n), ]
-  Y <- y.all[1:n]
-  Y.test <- y.all[-(1:n)]
-  
-  # about beta
-  non0.beta = as.vector(SBP.true != 0)
-  bspars = sum(non0.beta)
-  # solve for beta
-  c1plusc2 = theta.value * sum(abs(unique(ilrtrans.true$ilr.trans)))
-  beta.true = (b1 / (ilrtrans.true$const * c1plusc2)) * 
-    as.vector(ilrtrans.true$ilr.trans)
-  
-  saveRDS(list(
-    X = X, Y = Y, X.test = X.test, Y.test = Y.test, 
-    SBP.true = SBP.true, beta.true = beta.true, 
-    non0.beta = non0.beta
-  ),
-  paste0(output_dir, "/data", file.end))
+  if(file.exists(paste0(output_dir, "/data", file.end))){
+    data.tmp = readRDS(paste0(output_dir, "/data", file.end))
+    X = data.tmp$X
+    Y = data.tmp$Y
+    X.test = data.tmp$X.test
+    Y.test = data.tmp$Y.test
+    SBP.true = data.tmp$SBP.true
+    llc.coefs.true = data.tmp$llc.coefs.true
+    llc.coefs.non0 = data.tmp$llc.coefs.non0
+  } else{
+    # get latent variable
+    U.all = matrix(runif(min = -ulimit, max = ulimit, 2 * n), ncol = 1)
+    # simulate y from latent variable
+    y.all = as.vector(b0 + b1 * U.all + rnorm(2 * n) * sigma_y)
+    # simulate X: 
+    epsj.all = matrix(rnorm(2 * n * (p - 1)), nrow = (2 * n)) * sigma_x
+    a1 = theta.value * ilrtrans.true$ilr.trans[-p] 
+    #   alpha1j = {
+    #     c1=theta*ilr.const/k+   if j \in I+
+    #     -c2=-theta*ilr.const/k-  if j \in I-
+    #     0                       o/w
+    #   }
+    alrXj.all = a0 + U.all %*% t(a1) + epsj.all #log(Xj/Xp) =alpha0j+alpha1j*U+epsj
+    X.all <- alrinv(alrXj.all)
+    colnames(X.all) = paste0('s', 1:p)
+    
+    # subset out training and test sets
+    X = X.all[1:n, ]
+    X.test = X.all[-(1:n), ]
+    Y <- y.all[1:n]
+    Y.test <- y.all[-(1:n)]
+    
+    # about linear log-contrast models' coefficients
+    llc.coefs.non0 = as.vector(SBP.true != 0)
+    # solve for beta
+    c1plusc2 = theta.value * sum(abs(unique(ilrtrans.true$ilr.trans)))
+    llc.coefs.true = (b1 / (ilrtrans.true$const * c1plusc2)) * 
+      as.vector(ilrtrans.true$ilr.trans)
+    
+    saveRDS(list(
+      X = X, Y = Y, X.test = X.test, Y.test = Y.test, 
+      SBP.true = SBP.true, llc.coefs.true = llc.coefs.true, 
+      llc.coefs.non0 = llc.coefs.non0
+    ),
+    paste0(output_dir, "/data", file.end))
+  }
   
   ##############################################################################
   # compositional lasso
@@ -142,20 +152,21 @@ res = foreach(
   cl.lam.idx = which(classo$cvm <= oneSErule)[1]
   cl.a0 = classo$int[cl.lam.idx]
   cl.betahat = classo$bet[, cl.lam.idx]
-
+  
   # compute metrics on the selected model #
+  # prediction error
+  cl.Yhat.test = cl.a0 + log(X.test) %*% cl.betahat
+  cl.MSE.test = as.vector(crossprod(Y.test - cl.Yhat.test) / n)
+  # estimation accuracy, selection accuracy #
   cl.metrics = getMetricsLLC(
-    y.train = Y, y.test = Y.test,
-    logX.train = log(X),
-    logX.test = log(X.test),
-    n.train = n, n.test = n,
-    betahat0 = cl.a0, betahat = cl.betahat,
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true)
+    est.llc.coefs = cl.betahat,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true,
+    metrics = c("estimation", "selection"))
 
   saveRDS(c(
+    "mse" = cl.MSE.test,
     cl.metrics,
-    "betasparsity" = bspars,
     "logratios" = 0,
     "time" = cl.timing
   ),
@@ -180,7 +191,8 @@ res = foreach(
   slrspec = slr(
     x = X, y = Y, screen.method = "wald", cluster.method = "spectral",
     response.type = "continuous", s0.perc = 0, zeta = 0, 
-    threshold = slrspeccv$threshold[slrspeccv$index["1se",]])
+    threshold = slrspeccv$threshold[slrspeccv$index["1se",]], 
+    positive.slope = TRUE)
   end.time = Sys.time()
   slrspec.timing = difftime(
     time1 = end.time, time2 = start.time, units = "secs")
@@ -189,24 +201,24 @@ res = foreach(
   rownames(slrspec.fullSBP) = colnames(X)
   slrspec.fullSBP[match(
     names(slrspec$sbp), rownames(slrspec.fullSBP))] = slrspec$sbp
-  
   slrspec.coefs = getCoefsBM(
     coefs = coefficients(slrspec$fit), sbp = slrspec.fullSBP)
   
   # compute metrics on the selected model #
+  # prediction error
+  slrspec.Yhat.test = slrspec.coefs$a0 + 
+    slr.fromContrast(X.test, slrspec.fullSBP) * slrspec.coefs$bm.coefs
+  slrspec.MSE.test = as.vector(crossprod(Y.test - slrspec.Yhat.test) / n)
+  # estimation accuracy, selection accuracy #
   slrspec.metrics = getMetricsBM(
-    y.train = Y, y.test = Y.test,
-    ilrX.train = slr.fromContrast(X, sbp = slrspec.fullSBP),
-    ilrX.test = slr.fromContrast(X.test, sbp = slrspec.fullSBP),
-    n.train = n, n.test = n,
-    thetahat0 = slrspec.coefs$a0, thetahat = slrspec.coefs$bm.coefs,
-    betahat = slrspec.coefs$llc.coefs,
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true)
+    est.llc.coefs = slrspec.coefs$llc.coefs,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true,
+    metrics = c("estimation", "selection"))
   
   saveRDS(c(
+    "mse" = slrspec.MSE.test,
     slrspec.metrics,
-    "betasparsity" = bspars,
     "logratios" = sum(slrspec.coefs$bm.coefs != 0),
     "time" = slrspec.timing
   ),
@@ -215,7 +227,7 @@ res = foreach(
   ##############################################################################
   # slr
   #   screen.method = "wald"
-  #   cluster.method = "hieararchical"
+  #   cluster.method = "hierarchical"
   #   response.type = "continuous"
   #   s0.perc = 0
   #   zeta = 0
@@ -231,7 +243,8 @@ res = foreach(
   slrhier = slr(
     x = X, y = Y, screen.method = "wald", cluster.method = "hierarchical",
     response.type = "continuous", s0.perc = 0, zeta = 0, 
-    threshold = slrhiercv$threshold[slrhiercv$index["1se",]])
+    threshold = slrhiercv$threshold[slrhiercv$index["1se",]], 
+    positive.slope = TRUE)
   end.time = Sys.time()
   slrhier.timing = difftime(
     time1 = end.time, time2 = start.time, units = "secs")
@@ -240,24 +253,24 @@ res = foreach(
   rownames(slrhier.fullSBP) = colnames(X)
   slrhier.fullSBP[match(
     names(slrhier$sbp), rownames(slrhier.fullSBP))] = slrhier$sbp
-  
   slrhier.coefs = getCoefsBM(
     coefs = coefficients(slrhier$fit), sbp = slrhier.fullSBP)
   
   # compute metrics on the selected model #
+  # prediction error
+  slrhier.Yhat.test = slrhier.coefs$a0 + 
+    slr.fromContrast(X.test, slrhier.fullSBP) * slrhier.coefs$bm.coefs
+  slrhier.MSE.test = as.vector(crossprod(Y.test - slrhier.Yhat.test) / n)
+  # estimation accuracy, selection accuracy #
   slrhier.metrics = getMetricsBM(
-    y.train = Y, y.test = Y.test,
-    ilrX.train = slr.fromContrast(X, sbp = slrhier.fullSBP),
-    ilrX.test = slr.fromContrast(X.test, sbp = slrhier.fullSBP),
-    n.train = n, n.test = n,
-    thetahat0 = slrhier.coefs$a0, thetahat = slrhier.coefs$bm.coefs,
-    betahat = slrhier.coefs$llc.coefs,
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true)
+    est.llc.coefs = slrhier.coefs$llc.coefs,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true,
+    metrics = c("estimation", "selection"))
   
   saveRDS(c(
+    "mse" = slrhier.MSE.test,
     slrhier.metrics,
-    "betasparsity" = bspars,
     "logratios" = sum(slrhier.coefs$bm.coefs != 0),
     "time" = slrhier.timing
   ),
@@ -281,14 +294,7 @@ res = foreach(
     check = TRUE)
   
   # compute metrics on the selected model #
-  # prediction errors
-  # get prediction error on training set
-  slbl.Yhat.train = predict.glm(
-    slbl$glm, 
-    newdata = data.frame(V1 = balance::balance.fromSBP(
-      x = slbl.data$X, y = slbl.coefs$sbp)), 
-    type = "response")
-  slbl.PE.train = crossprod(Y - slbl.Yhat.train) / n
+  # prediction error
   # get prediction error on test set
   slbl.test.data = getSelbalData(X = X.test, y = Y.test, classification = FALSE)
   slbl.Yhat.test = predict.glm(
@@ -296,17 +302,17 @@ res = foreach(
     newdata = data.frame(V1 = balance::balance.fromSBP(
       x = slbl.test.data$X, y = slbl.coefs$sbp)), 
     type = "response")
-  slbl.PE.test = crossprod(Y.test - slbl.Yhat.test) / n
+  slbl.MSE.test = crossprod(Y.test - slbl.Yhat.test) / n
   # beta estimation accuracy, selection accuracy #
   slbl.metrics = getMetricsBM(
-    thetahat = slbl.coefs$bm.coefs, betahat = slbl.coefs$llc.coefs,
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true, metrics = c("betaestimation", "selection"))
-  slbl.metrics = c(PEtr = slbl.PE.train, PEte = slbl.PE.test, slbl.metrics)
+    est.llc.coefs = slbl.coefs$llc.coefs,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true, 
+    metrics = c("estimation", "selection"))
   
   saveRDS(c(
+    "mse" = slbl.MSE.test,
     slbl.metrics,
-    "betasparsity" = bspars,
     "logratios" = sum(slbl.coefs$bm.coefs != 0), 
     "time" = slbl.timing
   ),
@@ -350,10 +356,7 @@ res = foreach(
     codacore0.betahat = codacore0.coefs2$llc.coefs
     
     # compute metrics on the selected model #
-    # prediction errors
-    # get prediction error on training set
-    codacore0.Yhat.train = predict(codacore0, X)
-    # get prediction error on test set
+    # prediction error
     codacore0.Yhat.test = predict(codacore0, X.test)
     
   } else{
@@ -363,26 +366,20 @@ res = foreach(
     codacore0.betahat = rep(0, p)
     
     # compute metrics on the selected model #
-    # prediction errors
-    # get prediction error on training set
-    codacore0.Yhat.train = predict(codacore0model, X)
-    # get prediction error on test set
+    # prediction error
     codacore0.Yhat.test = predict(codacore0model, X.test)
   }
-  codacore0.PE.train = crossprod(Y - codacore0.Yhat.train) / n
-  codacore0.PE.test = crossprod(Y.test - codacore0.Yhat.test) / n
+  codacore0.MSE.test = crossprod(Y.test - codacore0.Yhat.test) / n
   
   # beta estimation accuracy, selection accuracy #
   codacore0.metrics = getMetricsBM(
-    betahat = codacore0.betahat,
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true, metrics = c("betaestimation", "selection"))
-  codacore0.metrics = c(
-    PEtr = codacore0.PE.train, PEte = codacore0.PE.test, codacore0.metrics)
+    est.llc.coefs = codacore0.betahat,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true, metrics = c("estimation", "selection"))
   
   saveRDS(c(
+    "mse" = codacore0.MSE.test,
     codacore0.metrics,
-    "betasparsity" = bspars,
     "logratios" = length(codacore0_coeffs), 
     "time" = codacore0.timing
   ),
@@ -404,27 +401,21 @@ res = foreach(
     time1 = end.time, time2 = start.time, units = "secs")
 
   # compute metrics on the selected model #
-  # prediction errors
-  # get prediction error on training set
-  lrl.Yhat.train = Wc %*% lrl$beta_min
-  lrl.PE.train = crossprod(Y - lrl.Yhat.train) / n
-  # get prediction error on test set
+  # prediction error
   Wc.test = scale(log(X.test), center = TRUE, scale = FALSE)
   Yc.test = Y.test - mean(Y.test)
   lrl.Yhat.test = Wc.test %*% lrl$beta_min
-  lrl.PE.test = crossprod(Yc.test - lrl.Yhat.test) / n
+  lrl.MSE.test = crossprod(Yc.test - lrl.Yhat.test) / n
 
   # beta estimation accuracy, selection accuracy #
   lrl.metrics = getMetricsBM(
-    betahat = lrl$beta_min, # don't back-scale bc only centered X (didn't scale)
-    true.sbp = SBP.true, non0.true.beta = non0.beta,
-    true.beta = beta.true, metrics = c("betaestimation", "selection"))
-  lrl.metrics = c(
-    PEtr = lrl.PE.train, PEte = lrl.PE.test, lrl.metrics)
+    est.llc.coefs = lrl$beta_min,
+    true.sbp = SBP.true, non0.true.llc.coefs = llc.coefs.non0,
+    true.llc.coefs = llc.coefs.true, metrics = c("estimation", "selection"))
 
   saveRDS(c(
+    "mse" = lrl.MSE.test,
     lrl.metrics,
-    "betasparsity" = bspars,
     "logratios" = NA,
     "time" = lrl.timing
   ),
